+++ atom/model_ops/attention_mla.py
@@ -209,41 +209,6 @@ class MLAAttention(nn.Module):
                 W_V, dtype=dtypes.fp8
             )
 
-            if is_plugin_mode() and is_vllm():
-                # The kernel operates on non-padded inputs. Hence, pre-compiling
-                # triton kernel to avoid runtime compilation for unseen batch sizes
-                # Pre-compile for batch sizes 1 to 1024 to cover most use-cases.
-                # On DS-R1, this step adds roughly 50s to the model loading time.
-                max_batch_size = 1024  # [ToDo] Find the optimal upper limit
-                pre_compilation_list = list(range(1, max_batch_size + 1))
-                from vllm.distributed.parallel_state import is_global_first_rank
-
-                if is_global_first_rank():
-                    pre_compilation_list = tqdm(
-                        pre_compilation_list,
-                        desc="[Aiter Triton] Pre-compiling fp8 BMM kernel",
-                        total=max_batch_size,
-                    )
-
-                for m in pre_compilation_list:
-                    x = torch.empty(
-                        (self.W_K.shape[0], m, self.W_K.shape[2]),
-                        dtype=torch.bfloat16,
-                        device=self.W_K.device,
-                    )
-                    x = _aiter_triton_fp8_bmm(
-                        x, self.W_K, self.W_K_scale, group_size=128, transpose_bm=True
-                    )
-
-                    x = torch.empty(
-                        (self.W_V.shape[0], m, self.W_V.shape[2]),
-                        dtype=torch.bfloat16,
-                        device=self.W_V.device,
-                    )
-                    x = _aiter_triton_fp8_bmm(
-                        x, self.W_V, self.W_V_scale, group_size=128, transpose_bm=True
-                    )
-
     @mark_trace(prefix="v_up_proj_and_o_proj", torch_compile=False)
     def _v_up_proj_and_o_proj(self, x):
         # Convert from (B, N, L) to (N, B, L)
